Fixed some syntax

+++ Lost_Chronicle/src/edu/auburn/eng/csse/COMP3710/Group_22/lost_chronicle/EventScheduler.java
@@ -78,8 +78,8 @@ public class EventScheduler {
 		event.setEnd_time(parseStringToDate(cursor.getString(3)));
 		event.setType(cursor.getString(4));
 		event.setOnGoing((short) Integer.parseInt(cursor.getString(5)));
-		event.setDifficulty((short) Integer.parseInt(cursor.getString(6)));
-		event.setEval((short) Integer.parseInt(cursor.getString(7)));
+		event.setDifficulty(Float.parseFloat(cursor.getString(6)));
+		event.setEval(Float.parseFloat(cursor.getString(7)));
 		event.setDescription(cursor.getString(8));
 		
 		//logcat it
